Renaming filenames correctly accoridng to standards
